+++ src/stores/quiz.ts
@@ -1,33 +1,28 @@
-// src/stores/quiz.ts
 import { defineStore } from 'pinia'
 import { ref } from 'vue'
 
 export interface Quiz {
-  quiz_id: string
-  quiz_set_id: number
+  quizId: string
+  quizSetId: number
   order: number
-  image_url: string
+  imageUrl: string
   answer: string
 }
 
 export const useQuizStore = defineStore('quiz', () => {
-  const quizSetId = ref<number | null>(null)
-  const quizzes = ref<Quiz[]>([])
+  const quizList = ref<Quiz[]>([])
 
-  function setQuizSet(id: number, data: Quiz[]) {
-    quizSetId.value = id
-    quizzes.value = data
+  function setQuizList(data: Quiz[]) {
+    quizList.value = data
   }
 
-  function clearQuizSet() {
-    quizSetId.value = null
-    quizzes.value = []
+  function clearQuizList() {
+    quizList.value = []
   }
 
   return {
-    quizSetId,
-    quizzes,
-    setQuizSet,
-    clearQuizSet,
+    quizList,
+    setQuizList,
+    clearQuizList,
   }
 })